# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMissionRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Documents",
            "type": "array"
          },
          "objective": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Objective",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9_-]{1,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "objective"
        ],
        "title": "CreateMissionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissionDetail": {
        "additionalProperties": false,
        "description": "A mission's current state, without the full event log.",
        "properties": {
          "clarification_question": {
            "default": "",
            "title": "Clarification Question",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "error_code": {
            "default": "",
            "title": "Error Code",
            "type": "string"
          },
          "error_message": {
            "default": "",
            "title": "Error Message",
            "type": "string"
          },
          "finding_count": {
            "default": 0,
            "title": "Finding Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "goal": {
            "default": "",
            "title": "Goal",
            "type": "string"
          },
          "has_report": {
            "default": false,
            "title": "Has Report",
            "type": "boolean"
          },
          "knowledge_claims": {
            "default": 0,
            "title": "Knowledge Claims",
            "type": "integer"
          },
          "knowledge_conflicts": {
            "default": 0,
            "title": "Knowledge Conflicts",
            "type": "integer"
          },
          "knowledge_entities": {
            "default": 0,
            "title": "Knowledge Entities",
            "type": "integer"
          },
          "knowledge_store": {
            "default": "",
            "title": "Knowledge Store",
            "type": "string"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "replan_iterations": {
            "default": 0,
            "title": "Replan Iterations",
            "type": "integer"
          },
          "required_operations": {
            "items": {
              "type": "string"
            },
            "title": "Required Operations",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "security_flags": {
            "items": {
              "$ref": "#/components/schemas/InjectionScan"
            },
            "title": "Security Flags",
            "type": "array"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "status": {
            "$ref": "#/components/schemas/MissionStatus"
          },
          "task_count": {
            "default": 0,
            "title": "Task Count",
            "type": "integer"
          },
          "termination_reason": {
            "default": "",
            "title": "Termination Reason",
            "type": "string"
          },
          "unresolved_gap_count": {
            "default": 0,
            "title": "Unresolved Gap Count",
            "type": "integer"
          },
          "verified_count": {
            "default": 0,
            "title": "Verified Count",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "objective",
          "status",
          "stage",
          "created_at"
        ],
        "title": "MissionDetail",
        "type": "object"
      },
      "MissionStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "CLARIFICATION_NEEDED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "MissionStatus",
        "type": "string"
      },
      "MissionSummary": {
        "additionalProperties": false,
        "description": "One row in the mission list.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "finding_count": {
            "default": 0,
            "title": "Finding Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "status": {
            "$ref": "#/components/schemas/MissionStatus"
          },
          "unresolved_gap_count": {
            "default": 0,
            "title": "Unresolved Gap Count",
            "type": "integer"
          },
          "verified_count": {
            "default": 0,
            "title": "Verified Count",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "objective",
          "status",
          "stage",
          "created_at"
        ],
        "title": "MissionSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "An adaptive, evidence-driven investigation engine.\n\nMissions run in the background: `POST /api/v1/missions` returns a run id immediately and\n`GET /api/v1/missions/{id}/stream` streams the run as it happens. Every finding carries the\nsource locators it rests on, and every claim the evidence does not support is reported as\nunsupported rather than dropped.",
    "title": "JARVIS",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/missions": {
      "get": {
        "operationId": "list_missions_api_v1_missions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MissionSummary"
                  },
                  "title": "Response List Missions Api V1 Missions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Missions",
        "tags": [
          "missions"
        ]
      },
      "post": {
        "description": "Start a mission. 202, not 201: the work is accepted, not finished.",
        "operationId": "create_mission_api_v1_missions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}": {
      "get": {
        "operationId": "get_mission_api_v1_missions__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/cancel": {
      "post": {
        "description": "Ask a mission to stop. 202: cancellation is cooperative, so it is requested, not done.\n\nThe run stops at its next await point - a tool call or model call boundary - and keeps\nwhatever it had already established.",
        "operationId": "cancel_mission_api_v1_missions__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/events": {
      "get": {
        "description": "The full event log, paginated. This is the run's reconstructable record.",
        "operationId": "get_events_api_v1_missions__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Events",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/findings": {
      "get": {
        "operationId": "get_findings_api_v1_missions__run_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Findings",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/v1/missions/{run_id}/gaps": {
      "get": {
        "operationId": "get_gaps_api_v1_missions__run_id__gaps_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GapRecord"
                  },
                  "title": "Response Get Gaps Api V1 Missions  Run Id  Gaps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Gaps",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/v1/missions/{run_id}/report": {
      "get": {
        "operationId": "get_report_api_v1_missions__run_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/report.md": {
      "get": {
        "description": "The same report a human reads, as Markdown.",
        "operationId": "get_report_markdown_api_v1_missions__run_id__report_md_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Markdown",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/stream": {
      "get": {
        "description": "Server-sent events for a running mission. See ADR-008.",
        "operationId": "stream_api_v1_missions__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{run_id}/tasks": {
      "get": {
        "description": "The task graph with live statuses, plus the waves that can run concurrently.",
        "operationId": "get_tasks_api_v1_missions__run_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskGraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tasks",
        "tags": [
          "missions"
        ]
      }
    },
...
  }
}